Fix debug text on some emulators

+++ tools/sotn-debugmode/debugmode.h
@@ -1,6 +1,6 @@
 #include <game.h>
 
 void InitFont();
-void SetFontCoord(int x, int y);
+void SetFontCoord(s16 x, s16 y);
 void BeginFont();
 void EndFont();
